Load the projects page at the root.
* Load the projects page at the root.
* Specify route paths in routes rather than in app.js.
* Rename features to project.
* Fix the rendering and update button on the project page.
