+++ lib/specifications/getProject.js
@@ -12,7 +12,7 @@ var featureFileRoot = path.join(__dirname, '..', '..', 'public', 'feature-files'
 /**
  * Get a copy of the project, derive metadata, store metadata.
  *
- * @return a promise for the completion of repo metadata storage.
+ * @return a promise for the project metadata on storage completion.
  */
 function getProject(repoUrl) {
   var repoName = /\/([^\/]+?)(?:\.git)?\/?$/.exec(repoUrl);
