+++ lib/specifications/projectDataStorage.js
@@ -12,12 +12,15 @@ var projectDataDirectory = path.join(appRoot, 'project-data');
 
 module.exports = {
 
-  // Return promise for write completion.
+  // Return promise for the metadata on write completion.
   persist: function(projectData) {
 
     // Ensure project data directory exists and write the file.
     fs.existsSync(projectDataDirectory) || fs.mkdirSync(projectDataDirectory);
-    return qfs.write(path.join(projectDataDirectory, projectData.name + '.data'), JSON.stringify(projectData));
+    return qfs.write(path.join(projectDataDirectory, projectData.name + '.data'), JSON.stringify(projectData))
+      .then(function() {
+        return projectData;
+      });
   },
 
   // Return promise for an array of paths in data directory
