+++ lib/specifications/projectMetaData.js
@@ -41,7 +41,7 @@ function deriveProjectData(featureFileRoot, repoData) {
     commit: repoData.commit,
     shortCommit: repoData.shortCommit,
     localName: repoData.localName,
-    projectLink: '/features#' + repoData.repoName,
+    projectLink: '/' + repoData.repoName,
     featureFilePaths: []
   }
 
@@ -59,7 +59,7 @@ function deriveProjectData(featureFileRoot, repoData) {
 /**
  * Return a function which takes the repo data, derives the project data and stores it.
  *
- * @return A promise for completion of the storage of the project data.
+ * @return A promise for the project data on completion of the storage.
  */
 function deriveAndStore(featureFileRoot) {
   return function(repoData) {
