--- views/features.hbs
@@ -1,30 +1,0 @@
-{{> layout_start_document}}
-<link rel="stylesheet" href="/public/css/features.css">
-{{> layout_start_body}}
-<main>
-  <header>
-    <h1>Available Specifications</h1>
-  </header>
-  {{#each projects}}
-    <section class="project">
-      <header>
-        <h2 class="project-title" id={{this.name}}>
-          <a class="quick-link" href="#{{this.name}}"><span class="quick-link-icon">&elsdot;</span><span class="project-name">{{this.name}}</span></a>
-        </h2>
-        <span class="project-commit">&commat;{{this.shortCommit}}</span>
-        <span class="repo-link"><a href="{{this.scmUrl}}">(visit repository)</a></span>
-        <button class="project-update call-to-action"><a href="?update={{this.name}}">Update</a></button>
-      </header>
-      <div class="links-container">
-        {{#each this.featureFilePaths}}
-          <p><a class="spec-link" href="{{this.featureRoute}}">{{this.featureName}}</a></p>
-        {{else}}
-          <p>Sorry, this project does not appear to contain any specifications</p>
-        {{/each}}
-      </div>
-    </section>
-  {{else}}
-    <p>Sorry, no projects were found</p>
-  {{/each}}
-</main>
-{{> layout_end}}
